orbital_backend/easysearchaws/requirements.txt
Flask==1.0.2
gunicorn==20.1.0
pycryptodome==3.9.6
Pygments==2.9.0
pyparsing==2.4.7
Pyrebase-t==1.0.28
requests==2.11.1
requests-toolbelt==0.9.1
flask-cors==3.0.10
Werkzeug==0.14.1
beautifulsoup4==4.9.3